requirements.txt
python-dotenv>=1.0.0
requests>=2.28.0
requests-oauthlib>=1.3.0
